Add sleeps to the test case for synced folders after manual VM reboot
First sleep is needed to make sure that ssh command will be long enough to be killed by reboot. Otherwise, it can exit fast with code 0, failing the assert.

The second sleep is needed to make sure that the subsequent command (cat) will be executed late enough, when the reboot is in progress. Otherwise, we can reach the guest before the actual reboot, which is not a valid verification.