Allow execute calls to receive extra env vars
With acceptance tests it can be useful to be able to override or set a
specific env var when executing a call at a single point in a test
example.

Instead of duplicating or moving with_temp_env() from the unit tests, it
is easier to ensure isolation of tests by avoiding modify of ENV
directly and instead allow the tests to configure the environment for
the process launched.